Admin page changes
Changed the admin control so that it is now a custom token connected to a UID instead of a firestore value (because it had some security risks to it, like users being able to create own POST req and changing adding isAdmin to their own UID). Also updated so that admins can add other admins.